Always allow user filetype extensions to override system config file
Before the user would sometimes have to override 'UnwantedFiletype=' in
the user config file to remove an extension written in the system file.
Geany would ignore an overridden filetype in the user config file that
also matches the extension. This happened when the wanted filetype had a
higher index than the unwanted one in filetypes_array.